requirements.txt
tkinter==8.6
pandas
reportlab
openpyxl
tabula-py
Pillow
tkcalendar
PyMuPDF==1.23.3
pyinstaller==6.3.0
